Move place model related files to app/models
