+++ app/models/place/place_model_helpers.rb
